+++ packages/react-dogfood/components/LayoutSelector.tsx
@@ -1,7 +1,5 @@
 import { Dispatch, SetStateAction, forwardRef, useEffect } from 'react';
 import {
-  CallParticipantsScreenView,
-  CallParticipantsView,
   GenericMenu,
   GenericMenuButtonItem,
   IconButton,
@@ -12,15 +10,21 @@ import {
   useHasOngoingScreenShare,
 } from '@stream-io/video-react-sdk';
 
+import {
+  CallParticipantsScreenView,
+  CallParticipantsView,
+  SpeakerOneOnOne,
+} from './CallLayout';
+
 export const LayoutMap = {
   LegacyGrid: {
     Component: CallParticipantsView,
-    title: 'Grid',
+    title: 'Grid (legacy)',
     props: {},
   },
   PaginatedGrid: {
     Component: PaginatedGridLayout,
-    title: 'Grid (beta)',
+    title: 'Paginated grid',
     props: {
       groupSize: 16,
     },
@@ -56,6 +60,11 @@ export const LayoutMap = {
   LegacySpeaker: {
     Component: CallParticipantsScreenView,
     title: 'Sidebar',
+    props: {},
+  },
+  SpeakerOneOnOne: {
+    Component: SpeakerOneOnOne,
+    title: 'Speaker 1:1',
     props: {},
   },
 };