+++ packages/react-native-dogfood/src/screens/Call/JoinCallScreen.tsx
@@ -9,8 +9,10 @@ import {
   View,
 } from 'react-native';
 import { useAppGlobalStoreValue } from '../../contexts/AppContext';
-import { useCall } from '@stream-io/video-react-native-sdk';
+import { useStreamVideoClient } from '@stream-io/video-react-native-sdk';
 import { MemberRequest } from '@stream-io/video-client';
+import { NavigationHeader } from '../../components/NavigationHeader';
+import { v4 as uuidv4 } from 'uuid';
 
 const styles = StyleSheet.create({
   container: {
@@ -76,7 +78,7 @@ const JoinCallScreen = () => {
   const [ringingUserIdsText, setRingingUserIdsText] = useState<string>('');
   const username = useAppGlobalStoreValue((store) => store.username);
   const [ringingUsers, setRingingUsers] = useState<string[]>([]);
-  const call = useCall();
+  const videoClient = useStreamVideoClient();
 
   const users = [
     { id: 'steve', name: 'Steve Galilli' },
@@ -87,15 +89,16 @@ const JoinCallScreen = () => {
   ];
 
   const startCallHandler = useCallback(async () => {
-    if (!call) {
-      return;
-    }
-
     let ringingUserIds = !ringingUserIdsText
       ? ringingUsers
       : ringingUserIdsText.split(',');
+
+    // we also need to add our own user id in the members
+    ringingUserIds = [...new Set([...ringingUserIds, username])];
+
     try {
-      await call.getOrCreate({
+      const call = videoClient?.call('default', uuidv4().toLowerCase());
+      await call?.getOrCreate({
         ring: true,
         data: {
           members: ringingUserIds.map<MemberRequest>((ringingUserId) => {
@@ -106,9 +109,9 @@ const JoinCallScreen = () => {
         },
       });
     } catch (error) {
-      console.log('Failed to createCall', call.id, 'default', error);
+      console.log('Failed to createCall', error);
     }
-  }, [call, ringingUserIdsText, ringingUsers]);
+  }, [ringingUserIdsText, ringingUsers, videoClient, username]);
 
   const isRingingUserSelected = (userId: string) =>
     ringingUsers.find((ringingUser) => ringingUser === userId);
@@ -125,6 +128,7 @@ const JoinCallScreen = () => {
 
   return (
     <SafeAreaView style={styles.container}>
+      <NavigationHeader />
       <View style={styles.textInputView}>
         <TextInput
           placeholder="Enter comma separated User Ids"