+++ packages/react-sdk/src/components/index.ts
@@ -10,7 +10,6 @@ export * from './Menu';
 export * from './Notification';
 export * from './PendingCallPanel';
 export * from './Permissions';
-export * from './StreamCall';
 export * from './StreamTheme';
 export * from './Search';
 export * from './Tooltip';