+++ packages/react-sdk/src/core/components/CallLayout/SpeakerLayout.tsx
@@ -31,7 +31,11 @@ import clsx from 'clsx';
 export type SpeakerLayoutProps = {
   ParticipantViewUISpotlight?: ParticipantViewProps['ParticipantViewUI'];
   ParticipantViewUIBar?: ParticipantViewProps['ParticipantViewUI'];
-  participantsBarPosition?: 'top' | 'bottom' | 'left' | 'right';
+  /**
+   * The position of the particpants who are not in focus.
+   * Providing `null` will hide the bar.
+   */
+  participantsBarPosition?: 'top' | 'bottom' | 'left' | 'right' | null;
 } & Pick<ParticipantViewProps, 'VideoPlaceholder'>;
 
 const DefaultParticipantViewUIBar = () => (
@@ -88,7 +92,8 @@ export const SpeakerLayout = ({
       <div
         className={clsx(
           'str-video__speaker-layout',
-          `str-video__speaker-layout--variant-${participantsBarPosition}`,
+          participantsBarPosition &&
+            `str-video__speaker-layout--variant-${participantsBarPosition}`,
         )}
       >
         <div className="str-video__speaker-layout__spotlight">
@@ -103,7 +108,7 @@ export const SpeakerLayout = ({
             />
           )}
         </div>
-        {otherParticipants.length > 0 && (
+        {otherParticipants.length > 0 && participantsBarPosition && (
           <div className="str-video__speaker-layout__participants-bar-buttons-wrapper">
             <div
               className="str-video__speaker-layout__participants-bar-wrapper"